+++ src/md5.inl
@@ -66,7 +66,7 @@ MD5_STATIC void md5_init(md5_state_t *pms);
 
 /* Append a string to the message. */
 MD5_STATIC void md5_append(md5_state_t *pms, const md5_byte_t *data,
-                           int nbytes);
+                           size_t nbytes);
 
 /* Finish the message and return the digest. */
 MD5_STATIC void md5_finish(md5_state_t *pms, md5_byte_t digest[16]);
@@ -403,10 +403,10 @@ MD5_STATIC void md5_init(md5_state_t *pms) {
 }
 
 MD5_STATIC void md5_append(md5_state_t *pms, const md5_byte_t *data,
-                           int nbytes) {
+                           size_t nbytes) {
 	const md5_byte_t *p = data;
-	int left = nbytes;
-	int offset = (pms->count[0] >> 3) & 63;
+	size_t left = nbytes;
+	size_t offset = (pms->count[0] >> 3) & 63;
 	md5_word_t nbits = (md5_word_t)(nbytes << 3);
 
 	if (nbytes <= 0)
@@ -420,7 +420,7 @@ MD5_STATIC void md5_append(md5_state_t *pms, const md5_byte_t *data,
 
 	/* Process an initial partial block. */
 	if (offset) {
-		int copy = (offset + nbytes > 64 ? 64 - offset : nbytes);
+		size_t copy = (offset + nbytes > 64 ? 64 - offset : nbytes);
 
 		memcpy(pms->buf + offset, p, copy);
 		if (offset + copy < 64)
